requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
google-generativeai>=0.7.2
supabase>=2.16.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
httpx>=0.27.0
requests>=2.32.0
